Fix issue #7
Clicking icon a second time blanks out the list - fixed by looking in DOM for "newArrayDiv" (which is the sidebar) and doing nothing if that exists.

It still opens popup.html and says "Injecting Script..." but it does not put another sidebar over the initial one and render the extension useless until reloading.  

Ultimately this isn't a 100% solution, but it addresses the problem adequately.